Build mapTagTree and give it to the Sidebar
This is rebased on top of master, to include recent tag bug fixes.

In TagStore, this creates a _mapTagTree object that structures the three-level
hierarchy of map navigation tags. `TagStore.getMapTagTree()` can be used to
retrieve this structure.

`App` stores the map tag tree in its state, and feeds it to Sidebar.

`Sidebar` iterates over the tag tree to create the nav with text.

Some additional small fixes:

+ MapStore uses `undefined` in place of `null` to check for missing data.
+ deleted console logging debug lines from `xlsx_2_db.js`

This commit _does not_ implement a nice nav for the map. It only makes sure
that the nav has the data it needs.